// src/app/Models/Usuarios.ts
export class Usuarios
{
    userid?: number;
    imagenuser?: any;
    rolid: number;
    typeid: number;
    nrodoc: string;
    idparametro: number;
    tarprof: string;
    name: string;
    lastname: string;
    birthdate: Date;
    phone: string;
    username: string;
    pass: string;
    email: string;
    datecreate: Date;
    dateexp: Date;
    ativeexp: boolean;
    active: boolean;
    superusersede: boolean;

    
}